requirements.txt
pytest
requests
PyYaml
jsonpath